#147 Asset type chooser tool
