+++ source/main/java/com/avereon/xenon/ToolManager.java
@@ -10,9 +10,7 @@ import com.avereon.xenon.asset.AssetEvent;
 import com.avereon.xenon.asset.AssetType;
 import com.avereon.xenon.task.Task;
 import com.avereon.xenon.task.TaskManager;
-import com.avereon.xenon.tool.ProgramTool;
-import com.avereon.xenon.tool.ToolInstanceMode;
-import com.avereon.xenon.tool.ToolRegistration;
+import com.avereon.xenon.throwable.NoToolRegisteredException;
 import com.avereon.xenon.workpane.Tool;
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
