+++ source/main/java/com/avereon/xenon/UiRegenerator.java
@@ -7,7 +7,6 @@ import com.avereon.xenon.asset.Asset;
 import com.avereon.xenon.asset.AssetType;
 import com.avereon.xenon.asset.OpenAssetRequest;
 import com.avereon.xenon.asset.type.ProgramWelcomeType;
-import com.avereon.xenon.tool.ProgramTool;
 import com.avereon.xenon.workpane.Tool;
 import com.avereon.xenon.workpane.Workpane;
 import com.avereon.xenon.workpane.WorkpaneEdge;
