+++ source/main/resources/com/avereon/xenon/bundles/asset.properties
@@ -24,3 +24,5 @@ notice-name=Notice
 notice-description=Product notice asset type
 fault-name=Fault
 fault-description=A program fault
+asset-name=New Asset
+asset-description=A program asset
