+++ source/main/resources/com/avereon/xenon/bundles/asset_es.properties
@@ -24,3 +24,5 @@ notice-name=Aviso
 notice-description=Typo de activo de aviso
 fault-name=Falla
 fault-description=Una falla del programa
+asset-name=Activo Nuevo
+asset-description=Un activo del programa
