+++ source/main/resources/com/avereon/xenon/bundles/tool.properties
@@ -29,4 +29,5 @@ product-available-refresh=Refreshing available products...
 product-installed-refresh=Refreshing installed products...
 product-updates-refresh=Checking for updates...
 fault-name=Fault
+asset-name=Asset
