+++ source/main/resources/com/avereon/xenon/bundles/tool_es.properties
@@ -29,3 +29,4 @@ product-available-refresh=Refrescando productos disponibles...
 product-installed-refresh=Refrescando productos instalados...
 product-updates-refresh=Comprobando actualizaciones...
 fault-name=Falla
+asset-name=Activo
