+++ source/test/java/com/avereon/xenon/ScreenShots.java
@@ -5,9 +5,9 @@ import com.avereon.util.OperatingSystem;
 import com.avereon.venza.javafx.FxUtil;
 import com.avereon.xenon.asset.type.ProgramAboutType;
 import com.avereon.xenon.asset.type.ProgramSettingsType;
-import com.avereon.xenon.tool.about.AboutTool;
+import com.avereon.xenon.tool.AboutTool;
 import com.avereon.xenon.tool.settings.SettingsTool;
-import com.avereon.xenon.tool.welcome.WelcomeTool;
+import com.avereon.xenon.tool.WelcomeTool;
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
 import com.avereon.xenon.workpane.WorkpaneEvent;
