+++ source/test/java/com/avereon/xenon/tool/AboutToolCloseAssetCloseToolUIT.java
@@ -1,32 +1,33 @@
-package com.avereon.xenon.tool.task;
+package com.avereon.xenon.tool;
 
-import com.avereon.xenon.asset.type.ProgramTaskType;
-import com.avereon.xenon.tool.ProgramTool;
+import com.avereon.xenon.asset.type.ProgramAboutType;
+import com.avereon.xenon.ProgramTool;
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
-import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.Future;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
+import static org.hamcrest.Matchers.is;
 
-class TaskToolCloseAssetCloseToolUIT extends TaskToolUIT {
+class AboutToolCloseAssetCloseToolUIT extends AboutToolUIT {
 
 	@Test
 	void execute() throws Exception {
 		Workpane pane = program.getWorkspaceManager().getActiveWorkspace().getActiveWorkarea().getWorkpane();
-		assertThat( pane.getTools().size(), Matchers.is( 0 ) );
+		assertThat( pane.getTools().size(), is( 0 ) );
 
-		Future<ProgramTool> future = program.getAssetManager().openAsset( ProgramTaskType.URI );
+		Future<ProgramTool> future = program.getAssetManager().openAsset( ProgramAboutType.URI );
 		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
-		assertThat( pane.getActiveTool(), instanceOf( TaskTool.class ) );
-		assertThat( pane.getTools().size(), Matchers.is( 1 ) );
+		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
+		assertThat( pane.getActiveTool(), instanceOf( AboutTool.class ) );
+		assertThat( pane.getTools().size(), is( 2 ) );
 
 		program.getAssetManager().closeAssets( future.get().getAsset() );
 		workpaneWatcher.waitForEvent( ToolEvent.REMOVED );
-		assertThat( pane.getTools().size(), Matchers.is( 0 ) );
+		assertThat( pane.getTools().size(), is( 1 ) );
 	}
 
 }
