+++ source/test/java/com/avereon/xenon/tool/AboutToolOpenTwiceUIT.java
@@ -1,33 +1,32 @@
-package com.avereon.xenon.tool.about;
+package com.avereon.xenon.tool;
 
-import com.avereon.xenon.asset.type.ProgramAboutType;
-import com.avereon.xenon.tool.ProgramTool;
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
 import org.junit.jupiter.api.Test;
-
-import java.util.concurrent.Future;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 
-class AboutToolCloseAssetCloseToolUIT extends AboutToolUIT {
+class AboutToolOpenTwiceUIT extends AboutToolUIT {
 
 	@Test
 	void execute() throws Exception {
 		Workpane pane = program.getWorkspaceManager().getActiveWorkspace().getActiveWorkarea().getWorkpane();
 		assertThat( pane.getTools().size(), is( 0 ) );
 
-		Future<ProgramTool> future = program.getAssetManager().openAsset( ProgramAboutType.URI );
+		clickOn( "#menu-help" );
+		clickOn( "#menuitem-about" );
 		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
 		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
 		assertThat( pane.getActiveTool(), instanceOf( AboutTool.class ) );
 		assertThat( pane.getTools().size(), is( 2 ) );
 
-		program.getAssetManager().closeAssets( future.get().getAsset() );
-		workpaneWatcher.waitForEvent( ToolEvent.REMOVED );
-		assertThat( pane.getTools().size(), is( 1 ) );
+		// Try to open the tool again and make sure there is still only one
+		clickOn( "#menu-help" );
+		clickOn( "#menuitem-about" );
+		workpaneWatcher.waitForEvent( ToolEvent.ACTIVATED );
+		assertThat( pane.getTools().size(), is( 2 ) );
 	}
 
 }
