+++ source/test/java/com/avereon/xenon/tool/AboutToolUIT.java
@@ -1,4 +1,4 @@
-package com.avereon.xenon.tool.about;
+package com.avereon.xenon.tool;
 
 import com.avereon.xenon.BaseToolUIT;
