+++ source/test/java/com/avereon/xenon/tool/TaskToolOpenTwiceUIT.java
@@ -1,32 +1,32 @@
-package com.avereon.xenon.tool.about;
+package com.avereon.xenon.tool;
 
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
-import static org.hamcrest.Matchers.is;
 
-class AboutToolOpenTwiceUIT extends AboutToolUIT {
+class TaskToolOpenTwiceUIT extends TaskToolUIT {
 
 	@Test
 	void execute() throws Exception {
 		Workpane pane = program.getWorkspaceManager().getActiveWorkspace().getActiveWorkarea().getWorkpane();
-		assertThat( pane.getTools().size(), is( 0 ) );
+		assertThat( pane.getTools().size(), Matchers.is( 0 ) );
 
 		clickOn( "#menu-help" );
-		clickOn( "#menuitem-about" );
+		clickOn( "#menuitem-task" );
 		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
-		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
-		assertThat( pane.getActiveTool(), instanceOf( AboutTool.class ) );
-		assertThat( pane.getTools().size(), is( 2 ) );
+		assertThat( pane.getActiveTool(), instanceOf( TaskTool.class ) );
+		assertThat( pane.getTools().size(), Matchers.is( 1 ) );
 
 		// Try to open the tool again and make sure there is still only one
 		clickOn( "#menu-help" );
-		clickOn( "#menuitem-about" );
+		clickOn( "#menuitem-task" );
 		workpaneWatcher.waitForEvent( ToolEvent.ACTIVATED );
-		assertThat( pane.getTools().size(), is( 2 ) );
+		assertThat( pane.getActiveTool(), instanceOf( TaskTool.class ) );
+		assertThat( pane.getTools().size(), Matchers.is( 1 ) );
 	}
 
 }
