+++ source/test/java/com/avereon/xenon/tool/TaskToolOpenUIT.java
@@ -1,4 +1,4 @@
-package com.avereon.xenon.tool.task;
+package com.avereon.xenon.tool;
 
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
 
-class TaskToolOpenTwiceUIT extends TaskToolUIT {
+class TaskToolOpenUIT extends TaskToolUIT {
 
 	@Test
 	void execute() throws Exception {
@@ -18,15 +18,9 @@ class TaskToolOpenTwiceUIT extends TaskToolUIT {
 		clickOn( "#menu-help" );
 		clickOn( "#menuitem-task" );
 		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
-		assertThat( pane.getActiveTool(), instanceOf( TaskTool.class ) );
-		assertThat( pane.getTools().size(), Matchers.is( 1 ) );
 
-		// Try to open the tool again and make sure there is still only one
-		clickOn( "#menu-help" );
-		clickOn( "#menuitem-task" );
-		workpaneWatcher.waitForEvent( ToolEvent.ACTIVATED );
-		assertThat( pane.getActiveTool(), instanceOf( TaskTool.class ) );
 		assertThat( pane.getTools().size(), Matchers.is( 1 ) );
+		assertThat( pane.getActiveTool(), instanceOf( TaskTool.class ) );
 	}
 
 }
