+++ source/test/java/com/avereon/xenon/tool/TaskToolUIT.java
@@ -1,4 +1,4 @@
-package com.avereon.xenon.tool.task;
+package com.avereon.xenon.tool;
 
 import com.avereon.xenon.BaseToolUIT;
