+++ source/test/java/com/avereon/xenon/tool/WelcomeToolCloseAssetCloseToolUIT.java
@@ -1,26 +1,33 @@
-package com.avereon.xenon.tool.welcome;
+package com.avereon.xenon.tool;
 
+import com.avereon.xenon.asset.type.ProgramWelcomeType;
+import com.avereon.xenon.ProgramTool;
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.instanceOf;
-import static org.hamcrest.Matchers.is;
+import java.util.concurrent.Future;
 
-class WelcomeToolOpenUIT extends WelcomeToolUIT {
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
+
+class WelcomeToolCloseAssetCloseToolUIT extends WelcomeToolUIT {
 
 	@Test
 	void execute() throws Exception {
 		Workpane pane = program.getWorkspaceManager().getActiveWorkspace().getActiveWorkarea().getWorkpane();
 		assertThat( pane.getTools().size(), is( 0 ) );
 
-		clickOn( "#menu-help" );
-		clickOn( "#menuitem-welcome" );
+		Future<ProgramTool> future = program.getAssetManager().openAsset( ProgramWelcomeType.URI );
 		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
 		assertThat( pane.getActiveTool(), instanceOf( WelcomeTool.class ) );
 		assertThat( pane.getActiveView().isMaximized(), is( true ) );
 		assertThat( pane.getTools().size(), is( 1 ) );
+
+		program.getAssetManager().closeAssets( future.get().getAsset() );
+		workpaneWatcher.waitForEvent( ToolEvent.REMOVED );
+		assertThat( pane.getMaximizedView(), is( nullValue() ) );
+		assertThat( pane.getTools().size(), is( 0 ) );
 	}
 
 }
