+++ source/test/java/com/avereon/xenon/tool/WelcomeToolOpenUIT.java
@@ -1,4 +1,4 @@
-package com.avereon.xenon.tool.welcome;
+package com.avereon.xenon.tool;
 
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
@@ -8,7 +8,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 
-class WelcomeToolOpenTwiceUIT extends WelcomeToolUIT {
+class WelcomeToolOpenUIT extends WelcomeToolUIT {
 
 	@Test
 	void execute() throws Exception {
@@ -18,13 +18,6 @@ class WelcomeToolOpenTwiceUIT extends WelcomeToolUIT {
 		clickOn( "#menu-help" );
 		clickOn( "#menuitem-welcome" );
 		workpaneWatcher.waitForEvent( ToolEvent.ADDED );
-		assertThat( pane.getActiveTool(), instanceOf( WelcomeTool.class ) );
-		assertThat( pane.getActiveView().isMaximized(), is( true ) );
-		assertThat( pane.getTools().size(), is( 1 ) );
-
-		clickOn( "#menu-help" );
-		clickOn( "#menuitem-welcome" );
-		workpaneWatcher.waitForEvent( ToolEvent.ACTIVATED );
 		assertThat( pane.getActiveTool(), instanceOf( WelcomeTool.class ) );
 		assertThat( pane.getActiveView().isMaximized(), is( true ) );
 		assertThat( pane.getTools().size(), is( 1 ) );
