+++ source/test/java/com/avereon/xenon/tool/guide/GuidedToolUIT.java
@@ -6,8 +6,8 @@ import com.avereon.xenon.Program;
 import com.avereon.xenon.ProgramProduct;
 import com.avereon.xenon.asset.Asset;
 import com.avereon.xenon.asset.MockAssetType;
-import com.avereon.xenon.tool.ToolInstanceMode;
-import com.avereon.xenon.tool.ToolRegistration;
+import com.avereon.xenon.ToolInstanceMode;
+import com.avereon.xenon.ToolRegistration;
 import com.avereon.xenon.workpane.ToolEvent;
 import javafx.application.Platform;
 import javafx.scene.control.TreeItem;
