+++ source/test/java/com/avereon/xenon/tool/settings/SettingsToolCloseAssetCloseToolUIT.java
@@ -1,7 +1,7 @@
 package com.avereon.xenon.tool.settings;
 
 import com.avereon.xenon.asset.type.ProgramSettingsType;
-import com.avereon.xenon.tool.ProgramTool;
+import com.avereon.xenon.ProgramTool;
 import com.avereon.xenon.workpane.ToolEvent;
 import com.avereon.xenon.workpane.Workpane;
 import org.junit.jupiter.api.Test;
